🔥🎨 creates util.db for general purpose functions; db.go for getting things from the db and returning
